Merge branch 'main' into ability-indicators

+++ Floarena/Assets/Constants/Data/Data.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 public static class Data
 {
     public static Item[] ITEMS = new Item[]{
-        new Item("Sturdy Soil", new Effect[]{new Effect(Attribute.HP, 100), new Effect(Attribute.MS, -0.2)}),
-        new Item("Light Soil", new Effect[]{new Effect(Attribute.HP, -20), new Effect(Attribute.MS, 0.4)}),
-        new Item("Prickly Thorns", new Effect[]{new Effect(Attribute.HP, 20), new Effect(Attribute.AD, 3)}),
+        new Item("Sturdy Soil", new Effect[]{new Effect(Attribute.HP, 100f), new Effect(Attribute.MS, -0.2f)}),
+        new Item("Light Soil", new Effect[]{new Effect(Attribute.HP, -20f), new Effect(Attribute.MS, 0.4f)}),
+        new Item("Prickly Thorns", new Effect[]{new Effect(Attribute.HP, 20f), new Effect(Attribute.AD, 3f)}),
     };
 
     public static Skill[] BASIC_SKILLS = new Skill[]{
@@ -46,7 +46,7 @@ public static class Data
     public static Skill[] DEFAULT_BASIC_SKILLS = new Skill[]{BASIC_SKILLS[0], BASIC_SKILLS[1]};
     public static Skill[] DEFAULT_ULTIMATE_SKILLS = new Skill[]{ULTIMATE_SKILLS[0]};
 
-    public static PlayerStats BASE_ATTRIBUTES = new PlayerStats(new Dictionary<Attribute, double>{
+    public static PlayerStats BASE_ATTRIBUTES = new PlayerStats(new Dictionary<Attribute, float>{
         { Attribute.HP, 100 },
         { Attribute.AD, 5 },
         { Attribute.AS, 1 },
@@ -54,7 +54,7 @@ public static class Data
         { Attribute.MS, 2 },
     });
 
-    public static PlayerStats ZEROED_ATTRIBUTES = new PlayerStats(new Dictionary<Attribute, double>{
+    public static PlayerStats ZEROED_ATTRIBUTES = new PlayerStats(new Dictionary<Attribute, float>{
         { Attribute.HP, 0 },
         { Attribute.AD, 0 },
         { Attribute.AS, 0 },
